requirements.txt
autorag
marker-pdf
pymupdf
streamlit
streamlit-option-menu
streamlit-extras
pandas
ollama
requests
pillow
tqdm
pyarrow
fastparquet
